+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 - Layout: Improve layout margins and typographical consistency across several content pages. (#5880, #5884, #5887, #5888, #5908)
 - Typography: Updated monospace font to Roboto Mono for consistency across login.gov sites. (#5891)
 - Icons: Replaced custom button icons using U.S. Web Design system icons. (#5904)
+- Analytics: Add ability to track when users go to Webauthn page to authenticate their device. (#5920)
 
 ### Accessibility